refactor(proxy): move gas_limit_calculator to neon-api
